Refactor code

+++ src/main/java/epam/com/gymapplication/service/TraineeService.java
@@ -1,6 +1,5 @@
 package epam.com.gymapplication.service;
 
-import epam.com.gymapplication.customexception.DaoException;
 import epam.com.gymapplication.customexception.ServiceException;
 import epam.com.gymapplication.dao.impl.TraineeDAOImpl;
 import epam.com.gymapplication.model.Trainee;
@@ -23,37 +22,73 @@ public class TraineeService {
 
 
     public void saveTrainee(Trainee trainee) throws ServiceException {
-        try {
-            traineeDAOImpl.save(trainee);
-            logger.info("Trainee saved successfully {}", trainee);
-        } catch (DaoException e) {
-            throw new ServiceException("There was a problem saving the Trainee");
+        if (trainee.getId() == null || trainee.getId() <= 0
+        || trainee.getUser().getId() == null || trainee.getUser().getId() <= 0
+                || trainee.getUser().getFirstName() == null
+                || trainee.getUser().getFirstName().isEmpty()
+                || trainee.getUser().getLastName() == null
+                || trainee.getUser().getLastName().isEmpty()
+                || trainee.getUser().getUsername() == null
+                || trainee.getUser().getUsername().isEmpty()
+                || trainee.getUser().getPassword() == null
+                || trainee.getUser().getPassword().isEmpty()
+                || trainee.getAddress() == null
+                || trainee.getAddress().isEmpty()
+                || trainee.getDateOfBirth() == null
+                || trainee.getDateOfBirth().toString().isEmpty()) {
+            logger.warn("Trainee save failed");
+            throw new ServiceException("Trainee save failed, trainee is invalid");
 
         }
+
+        traineeDAOImpl.save(trainee);
+        logger.info("Trainee saved {}", trainee);
+
     }
 
     public void updateTrainee(Trainee trainee) throws ServiceException {
-        try {
-            traineeDAOImpl.update(trainee);
-            logger.info("Trainee updated successfully {}", trainee);
-        } catch (DaoException e) {
-            throw new ServiceException("There was a problem updating the Trainee");
+        if (trainee.getId() == null || trainee.getId() <= 0
+                || trainee.getUser().getId() == null || trainee.getUser().getId() <= 0
+                || trainee.getUser().getFirstName() == null
+                || trainee.getUser().getFirstName().isEmpty()
+                || trainee.getUser().getLastName() == null
+                || trainee.getUser().getLastName().isEmpty()
+                || trainee.getUser().getUsername() == null
+                || trainee.getUser().getUsername().isEmpty()
+                || trainee.getUser().getPassword() == null
+                || trainee.getUser().getPassword().isEmpty()
+                || trainee.getAddress() == null
+                || trainee.getAddress().isEmpty()
+                || trainee.getDateOfBirth() == null
+                || trainee.getDateOfBirth().toString().isEmpty()) {
+            logger.warn("Trainee update failed");
+            throw new ServiceException("Trainee update failed, trainee is invalid");
         }
+
+        traineeDAOImpl.update(trainee);
+        logger.info("Trainee updated {}", trainee);
+
     }
 
     public void deleteTraineeById(Long id) throws ServiceException {
-        try {
-            traineeDAOImpl.deleteById(id);
-            logger.info("Trainee deleted successfully {}", id);
-        } catch (DaoException e) {
-            throw new ServiceException("There was a problem deleting Trainee by its id");
-
+        if (id == null || id <= 0) {
+            logger.warn("Trainee find by id failed");
+            throw new ServiceException("Trainee delete failed, id is invalid");
         }
+
+        traineeDAOImpl.deleteById(id);
+        logger.info("Trainee deleted {}", id);
+
     }
 
     public Optional<Trainee> findTraineeById(Long id)  {
-        logger.info("Trainee found by id {}", id);
-        return traineeDAOImpl.findById(id);
+        if (id == null || id <= 0) {
+            logger.warn("Trainee find id failed");
+            throw new ServiceException("Trainee find id failed, id is invalid");
+        }
+        logger.info("Found trainee by id {} ", id);
+        Optional<Trainee> byId = traineeDAOImpl.findById(id);
+        return Optional.of(byId.orElseThrow());
     }
 
     public Set<Map.Entry<Long, Trainee>> getAllTrainees() {
@@ -61,12 +96,20 @@ public class TraineeService {
     }
 
     public Optional<Trainee> findTraineeByFirstName(String firstName) {
+        if (firstName == null || firstName.isEmpty()) {
+            logger.warn("Trainee find by firstname failed");
+            throw new ServiceException("Trainee find by firstname find failed, firstName is invalid");
+        }
         logger.info("Trainee found by name {}", firstName);
-        return traineeDAOImpl.findByFirstName(firstName);
+        return Optional.of(traineeDAOImpl.findByFirstName(firstName).orElseThrow());
     }
 
     public Optional<Trainee> findByLastName(String lastName) {
+        if (lastName == null || lastName.isEmpty()) {
+            logger.warn("Trainee find by lastname  failed");
+            throw new ServiceException("Trainee find by lastname failed, lastName is invalid");
+        }
         logger.info("Trainee found by lastName {}", lastName);
-        return traineeDAOImpl.findByLastName(lastName);
+        return Optional.of(traineeDAOImpl.findByLastName(lastName).orElseThrow());
     }
 }
